add inputed date validation, check in date must lower than check out date

+++ features/reservations/models/reservation.go
@@ -11,10 +11,10 @@ type Reservation struct {
 	CheckInDate  time.Time `gorm:"not null;type:date"`
 	CheckOutDate time.Time `gorm:"not null;type:date"`
 	TotalNight   int       `gorm:"not null"`
-	TotalPrice   int       `gorm:"not null"`
+	TotalPrice   float64   `gorm:"not null"`
 	RoomID       uint
 	UserID       uint
 	RoomName     string
-	Price        int
-	Balance      int
+	Price        float64
+	Balance      float64
 }
